requirements.txt
python-telegram-bot
requests
python-dotenv
mem0ai
